fix category 삭제 수정

+++ mypage/views.py
@@ -5,6 +5,7 @@ from rest_framework.response import Response
 from .models import *
 from .serializers import *
 from rest_framework import status 
+from wish.models import Wish
 
 # Create your views here.
 
@@ -139,6 +140,10 @@ class CategoryView(views.APIView):
         category = Category.objects.filter(user=request.user, id=category_id).first()
         if not category:
             return Response({"error": "카테고리가 존재하지 않음"}, status=status.HTTP_404_NOT_FOUND)
+        category_obj = get_object_or_404(Category, id=category_id)
+        if Wish.objects.filter(category=category_obj).exists():
+          return Response({"error":"해당 카테고리에 위시아이템이 존재합니다."}, status=status.HTTP_400_BAD_REQUEST)
+
 
         # 카테고리 삭제
         category.delete()
